python_back_end/requirements.txt
# Core dependencies - numpy must be first for other packages
numpy #==1.26.0
setuptools
wheel
Cython

# Core FastAPI and web dependencies
fastapi
uvicorn[standard]
websockets
python-multipart
asyncpg
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
cryptography>=41.0.0
requests
pydantic
Pillow

# Audio processing - install torch first to avoid conflicts
torch==2.6.0
torchaudio==2.6.0
soundfile
openai-whisper
tavily-python

# TTS - our bread and butter!
chatterbox-tts
langchain
langchain-community
langchain-openai
langchain-postgres
sentence-transformers
ddgs
beautifulsoup4
newspaper3k
pytest
pytest-asyncio
pytest-mock
pytest-cov
httpx
responses
accelerate #==0.34.0
lxml_html_clean
python-dotenv
google-generativeai
psycopg2-binary
psycopg[binary]>=3.1.0
psycopg-pool>=3.1.0
pgvector>=0.2.0
docker>=7.0.0
defusedxml>=0.7.0

# Advanced research system dependencies
requests-cache>=1.0.0
trafilatura>=1.6.0
rank-bm25>=0.2.2

youtube-transcript-api>=0.6.0
yt-dlp>=2024.0.0
pypdf>=3.0.0
python-magic>=0.4.27
python-docx

# Document generation for artifacts
openpyxl>=3.1.0
python-pptx>=0.6.21
weasyprint>=60.0
reportlab>=4.0.0
